+++ client/src/components/Avatar.ts
