+++ client/src/components/AvatarDOM.tsx
